cluster: extract WaitForSegments to its own file
